# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LinkScope URL shortener and analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List the current user's short links, each with a stats summary
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkWithStats"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new short link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{id}:
    get:
      operationId: getLink
      tags: [links]
      summary: Get a single link with stats
      parameters:
        - $ref: "#/components/parameters/LinkId"
      responses:
        "200":
          description: Link detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLink
      tags: [links]
      summary: Update a link (destination, title, enabled state, expiration, preview settings)
      parameters:
        - $ref: "#/components/parameters/LinkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdate"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link
      parameters:
        - $ref: "#/components/parameters/LinkId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{id}/qrcode:
    get:
      operationId: getLinkQrCode
      tags: [links]
      summary: Get a QR code (as a data URL) for a link's short URL
      parameters:
        - $ref: "#/components/parameters/LinkId"
      responses:
        "200":
          description: QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    LinkId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Link:
      type: object
      required:
        - id
        - shortCode
        - shortUrl
        - originalUrl
        - title
        - isEnabled
        - expiresAt
        - previewTitle
        - previewDescription
        - previewImageUrl
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        shortCode:
          type: string
        shortUrl:
          type: string
          description: Full public short URL (clean /s/{code} path)
        legacyUrl:
          type: string
          description: Legacy /r/{code} redirect path, kept for backward compatibility
        originalUrl:
          type: string
        title:
          type: ["string", "null"]
        isEnabled:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        previewTitle:
          type: ["string", "null"]
        previewDescription:
          type: ["string", "null"]
        previewImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LinkStats:
      type: object
      required:
        - humanClicks
        - uniqueVisitors
        - botRequests
        - totalRequests
      properties:
        humanClicks:
          type: integer
        uniqueVisitors:
          type: integer
        botRequests:
          type: integer
        totalRequests:
          type: integer
    LinkWithStats:
      allOf:
        - $ref: "#/components/schemas/Link"
        - type: object
          required: [stats]
          properties:
            stats:
              $ref: "#/components/schemas/LinkStats"
    LinkInput:
      type: object
      required: [originalUrl]
      properties:
        originalUrl:
          type: string
          minLength: 1
        customAlias:
          type: string
        title:
          type: string
        expiresAt:
          type: ["string", "null"]
          format: date-time
        previewTitle:
          type: string
        previewDescription:
          type: string
        previewImageUrl:
          type: string
    LinkUpdate:
      type: object
      properties:
        originalUrl:
          type: string
          minLength: 1
        title:
          type: ["string", "null"]
        isEnabled:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        previewTitle:
          type: ["string", "null"]
        previewDescription:
          type: ["string", "null"]
        previewImageUrl:
          type: ["string", "null"]
    QrCode:
      type: object
      required: [dataUrl, shortUrl]
      properties:
        dataUrl:
          type: string
          description: base64 PNG data URL
        shortUrl:
          type: string
...