# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LinkScope URL shortener and analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: number;
  shortCode: string;
  /** Full public short URL (clean /s/{code} path) */
  shortUrl: string;
  /** Legacy /r/{code} redirect path, kept for backward compatibility */
  legacyUrl?: string;
  originalUrl: string;
  /** @nullable */
  title: string | null;
  isEnabled: boolean;
  /** @nullable */
  expiresAt: Date | null;
  /** @nullable */
  previewTitle: string | null;
  /** @nullable */
  previewDescription: string | null;
  /** @nullable */
  previewImageUrl: string | null;
  createdAt: Date;
  updatedAt: Date;
}